+++ quantforge/__init__.py
@@ -45,6 +45,7 @@ from .bates import bates_price, bates_smile
 from .double_heston import double_heston_price, double_heston_smile
 from .rough_heston import rough_heston_price, rough_heston_smile
 from .double_heston_calib import calibrate_double_heston
+from .heston_calib import calibrate_heston
 from .kou import kou_price, kou_smile
 from .cgmy import cgmy_price, cgmy_smile
 from .nig import nig_price, nig_smile
@@ -193,7 +194,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.129.0"
+__version__ = "1.130.0"
 
 __all__ = [
     "OptionType",
@@ -247,6 +248,7 @@ __all__ = [
     "rough_heston_price",
     "rough_heston_smile",
     "calibrate_double_heston",
+    "calibrate_heston",
     "kou_price",
     "kou_smile",
     "cgmy_price",